+++ packages/deposit-screener/src/screening/dummy.ts
@@ -2,6 +2,12 @@ import { Address } from "@nocturne-xyz/sdk";
 import { ScreeningApi } from ".";
 
 export class DummyScreeningApi implements ScreeningApi {
+  magicRejectionValue: bigint;
+
+  constructor(magicRejectionValue: number) {
+    this.magicRejectionValue = BigInt(magicRejectionValue);
+  }
+
   async isSafeDepositRequest(
     spender: Address,
     assetAddr: Address,
@@ -9,7 +15,11 @@ export class DummyScreeningApi implements ScreeningApi {
   ): Promise<boolean> {
     spender;
     assetAddr;
-    value;
+
+    if (value === this.magicRejectionValue) {
+      return false;
+    }
+
     return true;
   }
 }